requirements.txt
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
google-cloud-vision
openai
python-dotenv
pandas
Pillow
